+++ example/test/widget_test.dart
@@ -20,7 +20,7 @@
 //     expect(find.text('1'), findsNothing);
 
 //     // Tap the '+' icon and trigger a frame.
-//     await tester.tap(find.byIcon(Icons.add));
+//     await tester.tap(find.byIcon(Icons.add));`
 //     await tester.pump();
 
 //     // Verify that our counter has incremented.